refactor: remove collections manager and update documentation for admin UI management
